+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0
+
+- Verify signatures
+- Add Webauthn parser
+
 ## 1.0.0
 
 - First version